Moved types between files and made a few type fixes
