+++ gramhopper/configuration/boolean_operators.py
@@ -1,7 +1,7 @@
 import operator
 from boolean import boolean
 
-OPERATOR_TYPE_TO_FUNCTION = {
+OPERATOR_TO_FUNCTION = {
     boolean.AND: operator.and_,
     boolean.OR: operator.or_,
     boolean.NOT: operator.invert
