--- gramhopper/configuration/globals_dict.py
@@ -1,5 +1,0 @@
-from typing import Dict
-from .trigger_response import TriggerResponse
-
-
-GlobalsDict = Dict[str, TriggerResponse]
